Added new folders to project structure, moved most logic out of main class and moved it to sub classes to keep main cleaner.
