+++ Java/src/com/zeldaPlay/Player/ZeldaPlayerHelper.java
@@ -1,0 +1,45 @@
+package com.zeldaPlay.Player;
+
+import com.zeldaPlay.utils.Inputs;
+
+public class ZeldaPlayerHelper{
+
+  public static ZeldaPlayer makeZeldaPlayer() {
+    System.out.println("Please enter your name.");
+    String name = Inputs.getStringInput();
+    System.out.println("Great! Nice to meet you " + name + "! How strong are you?");
+    int str = Inputs.getIntInput(true);
+    System.out.println(str > 15 ? "Wow! That's pretty strong!" : "Oh, I'm guessing you rely on magic or something.");
+    System.out.println("How agile are you?");
+    int dex = Inputs.getIntInput(true);
+    System.out.println(dex > 15 ? "Very nice! I'm guessing you can't take a hit though."
+        : "Oh, well, I hope you can take a few hits on the field.");
+    System.out.println("Speaking of which, what is your constitution?");
+    int con = Inputs.getIntInput(true);
+    System.out.println(con > 15 ? "You can definitely take a hit!" : "Well, try not to get hit too much");
+    System.out.println("How smart would you say you are?");
+    int intel = Inputs.getIntInput(true);
+    System.out.println(intel > 15 ? "Oh don't we have a wiz kid over here!" : "Eh, intelligence is overrated anyways.");
+    System.out.println("What really matters is how sensible you are.");
+    int wis = Inputs.getIntInput(true);
+    System.out.println(wis > 15 ? "You definitely aren't stupid in the streets."
+        : "Maybe you shouldn't do the talking in a party. Don't want to offend anyone.");
+    System.out.println("Now, can you lie?");
+    int cha = Inputs.getIntInput(true);
+    System.out.println(cha > 15 ? "Remind me not to let you near my money. I'll never get it back"
+        : "Oh, good to know *snickers* sucker.");
+
+    ZeldaPlayer zeldaPlayer = new ZeldaPlayer(name, "", "Hylian", str, dex, con, intel, wis, cha);
+    System.out.println("For now, I'm assuming you're a Hylian, but that can change later");
+    System.out.println("Your name is:\t\t\t\t" + zeldaPlayer.getName());
+    System.out.println("Your strength is:\t\t\t" + zeldaPlayer.getStr() + "\t\t" + zeldaPlayer.getStrBon());
+    System.out.println("Your dexterity is:\t\t\t" + zeldaPlayer.getDex() + "\t\t" + zeldaPlayer.getDexBon());
+    System.out.println("Your constitution is:\t\t" + zeldaPlayer.getCon() + "\t\t" + zeldaPlayer.getConBon());
+    System.out.println("Your intelligence is:\t\t" + zeldaPlayer.getInt() + "\t\t" + zeldaPlayer.getIntBon());
+    System.out.println("Your wisdom is:\t\t\t\t" + zeldaPlayer.getWis() + "\t\t" + zeldaPlayer.getWisBon());
+    System.out.println("Your charisma is:\t\t\t" + zeldaPlayer.getCha() + "\t\t" + zeldaPlayer.getChaBon());
+    System.out.println("Your health points are:\t\t" + zeldaPlayer.getHealth());
+    System.out.println("Your magic points are:\t\t" + zeldaPlayer.getMagic());
+    return zeldaPlayer;
+  }
+}
