+++ README.md
@@ -1,6 +1,6 @@
 # zeldaPlay
 >A web application in progress for a home-brew D20 character tracker. Built in Java and Node.js and hosted on Heroku in separate instances
->The overall goal of this application is to strength my Java skills and to have a working application in the end. Maybe eventually take the Java application to an Android and iOS application that will be paired with the homebrew starter kit.
+>The overall goal of this application is to strength my Java skills and to have a working application in the end. Maybe eventually take the Java application to an Android and iOS application that will be paired with the home-brew starter kit.
 >
 >In the end, this application will be able to help track your character, make basic dice rolls for you, and allow you to digitally save multiple characters for ease of access.
 
@@ -34,7 +34,7 @@
 >
 > Those who are soldiers will have a bonus to using a sword and shield combo, and will be known to be on the front line of the battle, not wavering in the face of danger, and ready to fight for their kingdom and Goddess. 
 >
-> Those who are farm hands are more likely to be hospitable to those who visit their homes and are usually considered friendlier in general. These Hylians, while kind, are also known to be strong, durable, and very headstrong. Hylians of the farm hand background will be skilled in unarmed comabt, as herding goats can really be a workout. Along with the unarmed bonus, farm hands are also trained to use swords in the case of being called upon by the militia.
+> Those who are farm hands are more likely to be hospitable to those who visit their homes and are usually considered friendlier in general. These Hylians, while kind, are also known to be strong, durable, and very headstrong. Hylians of the farm hand background will be skilled in unarmed combat, as herding goats can really be a workout. Along with the unarmed bonus, farm hands are also trained to use swords in the case of being called upon by the militia.
 >
 > The few select Hylians who the Sheikah have deemed worthy of training are very withdrawn and reserved. They seem to always have their eyes out for those who are untrustworthy. These Hylians speak seldom and when they do, it is usually a warning. Those trained by the Sheikah are skilled in stealth combat, meaning they have a bonus when using daggers or swords.
 ### Goron
@@ -64,9 +64,9 @@
 >
 >
 ---
-## Java Verion
->A Java web application. Will be completed before the Node counterpart to have a stronger structure, more re-usable code, and better preformance (or so that's the idea).
+## Java Version
+>A Java web application. Will be completed before the Node counterpart to have a stronger structure, more re-usable code, and better performance (or so that's the idea).
 
-## Node Verion
+## Node Version
 >To be completed in the future.
